% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervention.list}
\alias{intervention.list}
\title{List of interventions in Germany}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 352 rows and 10 columns.
}
\source{
https://docs.google.com/spreadsheets/d/1cmGBMUhBt5y6jwiqaF7lh7VQNMN6D5FCoIltlOzhMfI/edit#gid=0
}
\usage{
intervention.list
}
\description{
Admin level, Bundesland, County, intervention type, start and end of implementation.
}
\keyword{datasets}
