% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interventions}
\alias{interventions}
\title{Dummy matrix for German interventions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7168 rows and 50 columns.
}
\usage{
interventions
}
\description{
Date, location (unit), Bundesland, and 46 rows for interventions. See by ls(interventions).
}
\keyword{datasets}
