% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regionaldatenbank}
\alias{regionaldatenbank}
\title{Population data by Regionaldatenbank}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2754 rows and 7 columns.
}
\source{
https://www.regionalstatistik.de/genesis/online/logon
Genesis Tabelle 12111-04-01-4
}
\usage{
regionaldatenbank
}
\description{
Name, admin level, age, total (population), area, and population density for county, state, and Germany.
}
\keyword{datasets}
