% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rki_new}
\alias{rki_new}
\title{Most recent case data based on Robert Koch Institute}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 217276 rows and 16 columns.
}
\source{
https://www.arcgis.com/sharing/rest/content/items/f10774f1c63e40168479a1feb6c7ca74/data
}
\usage{
rki_new
}
\description{
Data as known to the Robert Koch Institute at 2020-09-21.
}
\details{
Data is provided by Bundesamt für Kartographie und Geodäsie Robert Koch-Institut
under the open data license dl-de/by-2-0.
}
\examples{
# daily cases by reporting date (Meldedatum)
rki_new \%>\%
 group_by(Meldedatum) \%>\%
 summarise(cases = sum(AnzahlFall[Neuer.Fall \%in\% c(0,1)]))
}
\keyword{datasets}
