% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traced}
\alias{traced}
\title{Ratio of traced infectious cases}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100500 rows and 3 columns.
}
\usage{
traced
}
\description{
Date, location (unit), Bundesland, and 46 rows for interventions. See by ls(interventions).
}
\keyword{datasets}
