% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_dwd}
\alias{weather_dwd}
\title{Weather data from the German weather service}
\format{
An object of class \code{data.frame} with 220678 rows and 17 columns.
}
\source{
https://www.dwd.de/DE/klimaumwelt/cdc/cdc_node.html
}
\usage{
weather_dwd
}
\description{
Includes date, admin code and name for county (Landkreis), latitude, longitude,
and the weather variables
FM.Windgeschwindigkeit, FX.Windspitze, NM.Bedeckungsgrad,
PM.Luftdruck, RSK.Niederschlagshoehe, SDK.Sonnenscheindauer,
TGK.Lufttemperatur_5cm_min, TMK.Lufttemperatur,
TNK.Lufttemperatur_Min, TXK.Lufttemperatur_Max,
UPM.Relative_Feuchte und VPM.Dampfdruck.
}
\keyword{datasets}
